fixed a few critical issues spotted by Sonar: https://analysis.apache.org/drilldown/violations/37401?priority=CRITICAL
git-svn-id: https://svn.apache.org/repos/asf/camel/trunk@1137250 13f79535-47bb-0310-9956-ffa450edef68

+++ examples/camel-example-loan-broker/src/main/java/org/apache/camel/loanbroker/queue/version/Bank.java
@@ -44,7 +44,7 @@ public class Bank implements Processor {
         exchange.getOut().setBody("Bank processed the request.");
         // Sleep some time
         try {
-            Thread.sleep((int) (Math.random() * 10) * 100);
+            Thread.sleep((long) (Math.random() * 10) * 100);
         } catch (InterruptedException e) {
             // Discard
         }
